Tweaks

+++ Fluxanoia/src/main.hpp
@@ -13,8 +13,9 @@
 class Main {
 private:
 
-	const float tra_fac{ 0.05 };
-	const float rot_fac{ PI / 100.0f };
+	const float tra_fac{ 0.05f };
+	const float rot_fac{ 
+		static_cast<float>(PI) / 100.0f };
 
 	const int width;
 	const int height;
